Moved ImGui and tinyxml2 outside of Overload SDK
