+++ Sources/ImGui/imgui.cpp
@@ -818,13 +818,13 @@ CODE
 #define _CRT_SECURE_NO_WARNINGS
 #endif
 
-#include "OvUI/ImGui/imgui.h"
+#include "imgui.h"
 #ifndef IMGUI_DISABLE
 
 #ifndef IMGUI_DEFINE_MATH_OPERATORS
 #define IMGUI_DEFINE_MATH_OPERATORS
 #endif
-#include "OvUI/ImGui/imgui_internal.h"
+#include "imgui_internal.h"
 
 // System includes
 #include <ctype.h>      // toupper
