+++ Sources/ImGui/imgui_demo.cpp
@@ -66,7 +66,7 @@ Index of this file:
 #define _CRT_SECURE_NO_WARNINGS
 #endif
 
-#include "OvUI/ImGui/imgui.h"
+#include "imgui.h"
 #ifndef IMGUI_DISABLE
 
 #include <ctype.h>          // toupper
