+++ Sources/ImGui/imgui_draw.cpp
@@ -26,13 +26,13 @@ Index of this file:
 #define _CRT_SECURE_NO_WARNINGS
 #endif
 
-#include "OvUI/ImGui/imgui.h"
+#include "imgui.h"
 #ifndef IMGUI_DISABLE
 
 #ifndef IMGUI_DEFINE_MATH_OPERATORS
 #define IMGUI_DEFINE_MATH_OPERATORS
 #endif
-#include "OvUI/ImGui/imgui_internal.h"
+#include "imgui_internal.h"
 
 #include <stdio.h>      // vsnprintf, sscanf, printf
 #if !defined(alloca)
@@ -127,7 +127,7 @@ namespace IMGUI_STB_NAMESPACE
 #ifdef IMGUI_STB_RECT_PACK_FILENAME
 #include IMGUI_STB_RECT_PACK_FILENAME
 #else
-#include "OvUI/ImGui/imstb_rectpack.h"
+#include "imstb_rectpack.h"
 #endif
 #endif
 
@@ -150,7 +150,7 @@ namespace IMGUI_STB_NAMESPACE
 #ifdef IMGUI_STB_TRUETYPE_FILENAME
 #include IMGUI_STB_TRUETYPE_FILENAME
 #else
-#include "OvUI/ImGui/imstb_truetype.h"
+#include "imstb_truetype.h"
 #endif
 #endif
