+++ Sources/ImGui/imgui_impl_opengl3.cpp
@@ -58,8 +58,8 @@
 #define _CRT_SECURE_NO_WARNINGS
 #endif
 
-#include "OvUI/ImGui/imgui.h"
-#include "OvUI/ImGui/imgui_impl_opengl3.h"
+#include "imgui.h"
+#include "imgui_impl_opengl3.h"
 #include <stdio.h>
 #if defined(_MSC_VER) && _MSC_VER <= 1500 // MSVC 2008 or earlier
 #include <stddef.h>     // intptr_t
