+++ Sources/Overload/OvUI/include/OvUI/Widgets/AWidget.h
@@ -6,12 +6,13 @@
 
 #pragma once
 
+#include <imgui.h>
+
 #include <string>
 
 #include "OvUI/API/IDrawable.h"
 #include "OvUI/Plugins/Pluginable.h"
 #include "OvUI/Plugins/DataDispatcher.h"
-#include "OvUI/ImGui/imgui.h"
 
 namespace  OvUI::Internal { class WidgetContainer; }
