+++ Sources/Overload/OvUI/src/OvUI/Widgets/Buttons/Button.cpp
@@ -4,9 +4,11 @@
 * @licence: MIT
 */
 
-#include "OvUI/Widgets/Buttons/Button.h"
-#include "OvUI/Internal/Converter.h"
-#include "OvUI/ImGui/imgui_internal.h"
+#include <imgui.h>
+#include <imgui_internal.h>
+
+#include <OvUI/Widgets/Buttons/Button.h>
+#include <OvUI/Internal/Converter.h>
 
 OvUI::Widgets::Buttons::Button::Button(const std::string& p_label, const OvMaths::FVector2& p_size, bool p_disabled) :
 	label(p_label), size(p_size), disabled(p_disabled)
