+++ Sources/Overload/premake5.lua
@@ -24,18 +24,27 @@ dependdir = "%{wks.location}/../../Dependencies/"
 builddir = "%{wks.location}/../../Build/"
 resdir = "%{wks.location}/../../Resources/"
 
-include "OvAnalytics"
-include "OvAudio"
-include "OvCore"
-include "OvDebug"
-include "OvMaths"
-include "OvPhysics"
-include "OvRendering"
-include "OvTools"
-include "OvUI"
-include "OvWindowing"
+group "Dependencies"
+    include "../ImGui"
+    include "../tinyxml2"
+group ""
 
-include "OvEditor"
-include "OvGame"
+group "Overload SDK"
+    include "OvAnalytics"
+    include "OvAudio"
+    include "OvCore"
+    include "OvDebug"
+    include "OvMaths"
+    include "OvPhysics"
+    include "OvRendering"
+    include "OvTools"
+    include "OvUI"
+    include "OvWindowing"
+group ""
+
+group "Overload Applications"
+    include "OvEditor"
+    include "OvGame"
+group ""
 
 include "../../Resources"
